linspace now returns a lazy iterable of floats rather than a list.
there is also now a linspace_inclusive function that is similar to linspace but spreads the values a bit further out such that the iterator ends exactly at the  value rather than one step before.

there is now an addable protocol for types that can be added with __add__ to produce other objects of the same type.

there is now a cumul_sum function that returns a lazy iterable containing the cumulative sum of an iterable of addable elemeents.

there is now an iterable_as_collection function that takes an iterable that may or may not be a collection and returns the same iterable as a collection.  the returned collection may or may not be the original iterable.

there is now an is_approx function that takes two floats and an absolute tolerance and checks whether the absolute difference of those floats is less than or equal to the tolerance.